+++ benchmark/Benchmarks.hs
@@ -42,7 +42,7 @@ main = defaultMain
         ]
     , bgroup "Group operation"
         [ bench "rolling" $ nf (TS.rolling 20 sum) smallSeries
-        , bench "resampling" $ nf (TS.resample startTime (TS.seconds 20) sum) smallSeries
+        , bench "resampling" $ nf (TS.resample startTime (TS.seconds 20)) smallSeries
         ]
     , bgroup "IO"
         [ bench "load CSV" $ nfIO (TS.loadCSV TS.NoHeader parseISODateTime "testdata/test-100K.csv") ]